Fix AttributeError in richtext tile for different widget.
In a text package I am experimenting with using a different widget for the richtext field:
plone.app.form.widgets.wysiwygwidget.WYSIWYGWidget
That is what you would use in a portlet.
That returns a unicode instead of some data with an output and mimeType.
The transform fails then, so this commit just returns the unicode in that case.

I'm not yet completely sure this is sane, but it seems okay for now. :-)