Show the tile type on the compose tab:
This shows the type of a tile (Basic Tile, Collection Tile, etc...) in the
top right of the tile on the compose tab.  This makes it easier for the user
to know how they can add content -- especially if the user does not have
enough permission to view the layout tab.