Fix showing original size of images in banner and carousel tile.
Calling `scales.scale('image', scale)` when scale is `_original` would
give an empty image.  So we call it with `scale=None` in that case.

For Archetypes images this gives a BlobWrapper as result, which can
function mostly like a scale, except that it lacks the `url`
attribute.  In that case, for the src of the image we fall back to the
`download` page of the image item.