Fix test_populate_with_collection when tox is used
For some reason, when tox is used, the attributes of the tiles tags are
rendered in alphabetical order. When not using tox, the attributes are
rendered in the order they were defined in the pt file. So we use
lxml.etree to get the src attribute directly instead of comparing the
generated html.