Fix calendar tile in Plone 5.2
In Python calendar lib, the days of the week are values from 0 to 6. We
cannot use values outside this range. The `old_getEventsForCalendar`
method code was adding +1 to the first_weekday. So when the
`plone.first_weekday` record was 6, the first_weekday became 7, outside
the range 0 to 6, causing the error:

calendar.IllegalWeekdayError: bad weekday number 7; must be 0 (Monday)
to 6 (Sunday)

Now we don't add +1 to first_weekday.