Handles tile editing modal in Plone 5
We wait for the tile edit modal to be visible, before trying to interact
with it. The keyword "Click Edit Cover" was created to centralize the
click settings and wait  for the modal to be visible.

Also updated the locator of the Save and Cancel buttons. The keyword
"Click Button Save" doesn't find the button in the Plone 5 modal.