Fix test collection query
The index used by the collections query is the portal_type and not the
Type. See:

https://github.com/plone/plone.app.querystring/blob/1.4.15/plone/app/querystring/profiles/default/registry.xml#L467

And the portal_type index operation is
plone.app.querystring.operation.selection.any. See:

https://github.com/plone/plone.app.querystring/blob/1.4.15/plone/app/querystring/profiles/default/registry.xml#L473

When we create a collection programmatically, with a query with the
index Type, we get the js error:

Uncaught TypeError: o.indexes[e] is undefined

when we try to edit the collection through the web interface.