Fix WrongContainedType when editing a carousel tile
The problem was that the key "order" of the uuids field was coming in as
an int, when it should actually be a string. To see:

https://github.com/collective/collective.cover/blob/cb13a1c9f7094c237f697c1e1a81bf35ce39feeb/src/collective/cover/tiles/list.py#L37

In the migration to Python 3, the order field assignment was changed
from

six.text_type(order)

to

safe_unicode(order)

To see:

https://github.com/collective/collective.cover/commit/82c873b5bedd5a315149faf4112c244bbafe018a#diff-5736af52418c00117a6e6820d0dad4c838000e46066e775a0811421a2f10ac90L211-R2

However, `safe_unicode(order)` does not transform an int into a string,
like `six.text_type(order)`.

Then we use `six.text_type(order)` in the order field assignment