Fix tile html update after dragging and dropping content between tiles
The way js events were registered, when we dragged an image from a list
tile to a banner tile, two ajax calls were fired at the same time. One
for the `@@movetilecontent` view and one for the
`@@updatelisttilecontent` view. As both were fired at the same time and
both update tile data, a Conflict Error was occurring in
`@@updatelisttilecontent` in Plone 5.2. This caused the return of
`@@updatelisttilecontent` to take 2s. With this delay, the move_callback
function, which is responsible for removing the image from the list
tile, was executed before the html returned by `@@updatelisttilecontent`
was rendered. In order for move_callback to remove the image, it depends
on the html be rendered. As it wasn't, at the time of execution, the
image was not removed and was in the two tiles.

To fix the problem, I left the call of `@@updatelisttilecontent` in the
same event as the call of `@@movetilecontent`, and chaining the events
for move_callback to only be called after the html of
`@@updatelisttilecontent` is rendered.