Fix warning popup test when trying to exit layout editing without saving
The default Selenium profile for Firefox has the setting
"dom.disable_beforeunload" = True. So the onbeforeunload event popup,
which warns when leaving the layout tab without saving, is not being
shown, causing the test robot to fail.

It was necessary to create a custom Firefox profile, setting
"dom.disable_beforeunload" = False, so that the popup was displayed.

The keyword "Handle Alert" was also used instead of "Choose Cancel On
Next Confirmation", since "Choose Cancel On Next Confirmation" is
deprecated.