Unlock Cover when editor leaves Compose/Layout tabs
When an Editor enters the Compose/Layout tabs, the Cover is locked. But
when the Editor exit the Compose/Layout tabs the cover was not being
unlocked for the other Editors. Now the cover is unlocked.

In the Compose/Layout tabs, a request is sent to Plone to unlock the
cover, in the javascript unload event, which happens when the user
leaves the page.

The `plone.UnlockHandler.execute` function, which was called previously,
no longer exists in Plone 5.