Fix tile selection in Layout tab when we have multiple tiles
When we had multiple tiles or when the screen resolution was small, the
tile list was below the portlet column when there were portlets on the
right. This made it impossible to select tiles from the end of the list.

With this style change, the tile list is broken into two lines, without
going below the portlet column.