Fix test tile sorting
In commit

https://github.com/collective/collective.cover/commit/c403cbd8dbdd87d2b1b48aa27ee4e16e92a3a3d2

the funcition onStop of ComposeView, has been removed to fix the drag
and drop of content from one tile to another. However, it is still
needed to save the position of a content, when drag and drop occurs on
the same tile, for sorting within the tile.But we do a test, to make
sure it will only do something, if the drag and drop is done on the same
tile. Otherwise, an error occurs when moving content from one tile to
another.