Wait after saving cover to avoid locking error when trying to delete content
We are updating the content and will soon save and delete. The time
between saving and deleting was short, causing Plone to assume that the
content was still locked when trying to delete. This caused the error in
Plone:

   ResourceLockedError: Object 'title-1' is locked.

This caused the following error in test robot:

  Test Cover
  ==============================================================================
  Test CRUD
| FAIL |
  Page should have contained text 'has been deleted' but did not.

So now wait 2 seconds before deleting the content.