Fix error when clearing content chooser input
When we cleared the filter, the contents of the previous filter and the
amount of items found were still displayed.

This was due to a Javascript error. The variable portal_url was not
defined inside the function that is called when clicking on the clear
button. Then we use this.portal_url, to use the class variable. For this
to be considered the class, it was necessary to use arrow functions.