Adds icons for the Cover content type and for the Compose and Layout tabs
A font was created with the icons.

The viewlet that inserts the CSS generated by webpack, is inserted only
in the Cover content type. So, it was necessary to generate a different
CSS file, for the fonts, which is inserted in all types of content. This
was necessary, because the Cover type icon can appear in the folder
contents of any folder.